# Groot reader perf