a set of various optimisations that bring:

```
$> benchstat ref.txt new.txt
name                      old time/op    new time/op    delta
ReadTreeF64-4                483µs ± 3%     437µs ± 5%   -9.58%  (p=0.000 n=19+20)
ReadTreeSliF64/0-4           213µs ± 2%     154µs ± 2%  -27.57%  (p=0.000 n=20+19)
ReadTreeSliF64/1-4           213µs ± 1%     154µs ± 3%  -27.87%  (p=0.000 n=20+18)
ReadTreeSliF64/2-4           214µs ± 3%     156µs ± 4%  -27.31%  (p=0.000 n=20+20)
ReadTreeSliF64/4-4           214µs ± 3%     156µs ± 6%  -27.11%  (p=0.000 n=19+20)
ReadTreeSliF64/8-4           214µs ± 2%     155µs ± 5%  -27.50%  (p=0.000 n=18+20)
ReadTreeSliF64/16-4          214µs ± 3%     155µs ± 4%  -27.66%  (p=0.000 n=20+20)
ReadTreeSliF64/64-4          213µs ± 2%     155µs ± 3%  -27.14%  (p=0.000 n=20+20)
ReadTreeSliF64/128-4         216µs ± 4%     155µs ± 2%  -28.25%  (p=0.000 n=20+20)
ReadTreeSliF64/512-4         213µs ± 2%     157µs ± 3%  -26.34%  (p=0.000 n=19+19)
ReadTreeSliF64/1024-4        213µs ± 1%     155µs ± 1%  -27.43%  (p=0.000 n=17+16)
ReadTreeSliF64/1048576-4     213µs ± 2%     156µs ± 3%  -26.91%  (p=0.000 n=19+19)

name                      old alloc/op   new alloc/op   delta
ReadTreeF64-4                27.0B ± 0%     24.4B ± 6%   -9.47%  (p=0.000 n=15+18)
ReadTreeSliF64/0-4           10.0B ± 0%      7.3B ±10%  -27.00%  (p=0.000 n=14+20)
ReadTreeSliF64/1-4           10.0B ± 0%      7.0B ± 0%  -30.00%  (p=0.000 n=14+19)
ReadTreeSliF64/2-4           9.74B ± 8%     7.00B ± 0%  -28.11%  (p=0.000 n=19+20)
ReadTreeSliF64/4-4           10.0B ± 0%      7.0B ± 0%  -30.00%  (p=0.000 n=16+18)
ReadTreeSliF64/8-4           10.0B ± 0%      7.0B ± 0%  -30.00%  (p=0.000 n=14+19)
ReadTreeSliF64/16-4          10.0B ± 0%      7.0B ± 0%  -30.00%  (p=0.000 n=15+18)
ReadTreeSliF64/64-4          9.74B ± 8%     7.00B ± 0%  -28.11%  (p=0.000 n=19+18)
ReadTreeSliF64/128-4         10.0B ± 0%      7.0B ± 0%  -30.00%  (p=0.000 n=15+19)
ReadTreeSliF64/512-4         10.0B ± 0%      7.0B ± 0%  -30.00%  (p=0.000 n=16+19)
ReadTreeSliF64/1024-4        9.69B ± 7%     7.00B ± 0%  -27.74%  (p=0.000 n=16+17)
ReadTreeSliF64/1048576-4     10.0B ± 0%      7.0B ± 0%  -30.00%  (p=0.000 n=13+18)

name                      old allocs/op  new allocs/op  delta
ReadTreeF64-4                 0.00           0.00          ~     (all equal)
ReadTreeSliF64/0-4            0.00           0.00          ~     (all equal)
ReadTreeSliF64/1-4            0.00           0.00          ~     (all equal)
ReadTreeSliF64/2-4            0.00           0.00          ~     (all equal)
ReadTreeSliF64/4-4            0.00           0.00          ~     (all equal)
ReadTreeSliF64/8-4            0.00           0.00          ~     (all equal)
ReadTreeSliF64/16-4           0.00           0.00          ~     (all equal)
ReadTreeSliF64/64-4           0.00           0.00          ~     (all equal)
ReadTreeSliF64/128-4          0.00           0.00          ~     (all equal)
ReadTreeSliF64/512-4          0.00           0.00          ~     (all equal)
ReadTreeSliF64/1024-4         0.00           0.00          ~     (all equal)
ReadTreeSliF64/1048576-4      0.00           0.00          ~     (all equal)
```